+++ resources/assets/js/native.js
@@ -3,7 +3,9 @@ import IndexSorting from './utilities/sortable';
 /**
  * List here all the js utilities needed to be loaded after the Vue instantiation
  */
-require('./utilities/form-submit');
+import FormSubmit from './utilities/form-submit';
+
+FormSubmit.listen('[data-submit-form]');
 
 /**
  * Sortable