+++ resources/assets/js/utilities/EventBus.js
@@ -20,7 +20,7 @@ function subscribe(event, callback) {
     if (!subscriptions[event]) subscriptions[event] = {};
 
     subscriptions[event][id] = callback;
-    console.dir(subscriptions);
+
     return {
         unsubscribe: () => {
             delete subscriptions[event][id];