+++ resources/views/OLD/managers-old/_transitions/archive.blade.php
@@ -1,5 +1,5 @@
 @if($manager->isAssistedBy('archive'))
-    @include('chief::manager._modals.archive-modal')
+    @include('chief::manager._transitions.modals.archive-modal')
 
     <a v-cloak @click="showModal('archive-manager-<?= \Illuminate\Support\Str::slug($manager->assistant('archive')->route('archive')); ?>')" class="block p-3 text-warning --link-with-bg">
         Archiveer