+++ resources/views/manager/_transitions/panel/unpublish.blade.php
@@ -1,0 +1,20 @@
+<div class="flex justify-between items-center">
+    @if($linkForm->isAnyLinkOnline())
+        <span>De pagina staat online. 👍</span>
+
+        <form class="mb-0" action="@adminRoute('unpublish', $model)" method="POST">
+            {{ csrf_field() }}
+            <button type="submit" class="btn btn-error cursor-pointer">Haal offline</button>
+        </form>
+    @else
+        <span class="mr-4">De pagina staat gepubliceerd maar zal zonder link nog niet bereikbaar zijn. Voeg hieronder nog een link toe!</span>
+
+        <form class="mb-0" action="@adminRoute('unpublish', $model)" method="POST">
+            {{ csrf_field() }}
+            <button type="submit" class="btn btn-error cursor-pointer">Zet terug in draft</button>
+        </form>
+    @endif
+
+
+</div>
+